Combined image with artwork
- combined image and artwork models
- combined image and artwork routes
- adjusted server.js
