+++ app/models/image.js
@@ -1,15 +1,15 @@
-const mongoose = require('mongoose')
-
-const imageSchema = new mongoose.Schema({
-  imageName: {
-    type: String,
-    default: 'none',
-    required: true
-  },
-  imageData: {
-    type: String,
-    required: true
-  }
-})
-
-module.exports = mongoose.model('Image', imageSchema)
+// const mongoose = require('mongoose')
+//
+// const imageSchema = new mongoose.Schema({
+//   imageName: {
+//     type: String,
+//     default: 'none',
+//     required: true
+//   },
+//   imageData: {
+//     type: String,
+//     required: true
+//   }
+// })
+//
+// module.exports = mongoose.model('Image', imageSchema)
